refactor(FloatingButton): 상수 분리
