+++ frontend/src/components/common/FloatingButton/constants.ts
@@ -1,0 +1,14 @@
+import { ROUTE_PATHS_MAP } from "@constants/route";
+
+const SUB_BUTTONS = [
+  {
+    text: "✈️ 여행 계획 작성",
+    route: ROUTE_PATHS_MAP.travelPlanRegister,
+  },
+  {
+    text: "📝 여행기 작성",
+    route: ROUTE_PATHS_MAP.travelogueRegister,
+  },
+];
+
+export default SUB_BUTTONS;
